requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
python-dotenv
beautifulsoup4
sentence-transformers